requirements/base.txt
aiohttp == 2.0.7
uvloop == 0.8.0
cchardet == 2.0.0 # faster replacement for chardet, a requirement of aiohttp
python-dotenv == 0.6.4
aioredis == 0.3.0
faker == 0.7.11
arrow == 0.10.0
flask == 0.12.1
